Rename `Apply.h` to `Transform.h`
Tags: [refactor]
Compile: yes
Tests: yes
